+++ Gec_Xpress/admin/Edit_news_type.php
@@ -25,7 +25,7 @@ if(isset($_REQUEST["btnsave"])){
             <?php
     }
 }
-if(isset($_REQUEST["ntype_id"])){
+if(isset($_REQUEST["ntype_id"]) && $_REQUEST["false"]=='false'){
     //echo $_REQUEST["branch_id"];
     $qry_b="DELETE FROM news_type WHERE nid=".$_REQUEST["ntype_id"];
     if(mysqli_query($con, $qry_b))
@@ -44,6 +44,15 @@ if(isset($_REQUEST["ntype_id"])){
             alert(res);
         </script>
             <?php
+    }
+}
+if(isset($_REQUEST["ntype_id"]) && isset($_REQUEST["trash"]) && isset($_REQUEST["trash"]) == 'yes'){
+    $qry_h="UPDATE news_type SET trash='yes' WHERE nid=".$_REQUEST["ntype_id"];
+    if(mysqli_query($con, $qry_h)){
+        //echo "vikash";
+        ?><script>alert('News_type is moved to trash!!');
+                    window.location.href="Edit_news_type.php";
+        </script><?php
     }
 }
 ?>
@@ -70,7 +79,7 @@ if(isset($_REQUEST["ntype_id"])){
             </script>
     </head>
     <body>
-        <?php include './header.php'; ?>
+        <?phpad include './header.php'; ?>
     <div class="wrapper row3 center"  style="">
         <main class="hoc container clear"> 
           <!-- main body -->
@@ -90,13 +99,17 @@ if(isset($_REQUEST["ntype_id"])){
     
         <div class="container" style="margin-top: 20px;">
         <?php
-            $qry="SELECT * FROM news_type ORDER BY nid";
+        if(isset($_REQUEST["trash"]) && isset($_REQUEST["trash"]) == 'yes'){
+            $qry="SELECT * FROM news_type WHERE trash='yes' ORDER BY nid";
+        }else{
+            $qry="SELECT * FROM news_type WHERE trash!='yes' ORDER BY nid";
+        }
             $result= mysqli_query($con, $qry);
             while($row= mysqli_fetch_array($result))
             {
                 ?>                    
 		<div class="alert alert-success alert-dismissable fade in">
-                    <a href="Edit_news_type.php?ntype_id=<?php echo $row["nid"] ?>" class="close" data-dismiss="alert" area-label="close" title="Remove the branch" onClick='return confirmDelete(this);'>&times</a>
+                    <a href="Edit_news_type.php?ntype_id=<?php echo $row["nid"] ?> && trash='yes'" class="close" data-dismiss="alert" area-label="close" title="Remove the branch" onClick='return confirmDelete(this);'>&times</a>
 			<?php echo $row["ntype"]; ?>
 		</div>
                 <?php
